Modify(#86): place 조회할 때 rating 값도 넘겨주기

+++ src/main/java/shop/babsim/babsim/review/api/dto/response/ReviewInfoResDto.java
@@ -10,7 +10,7 @@ import shop.babsim.babsim.review.domain.Review;
 @Builder
 public record ReviewInfoResDto(
         List<String> feedImageUrls,
-        int rating,
+        double rating,
         String content,
         int likes,
         Long memberId,
